# Reduce the amount of slicing in GetMemory/GetSpan/Advance